# backend/services/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Source:
    properties:
      config:
        items:
          type: integer
        type: array
      created_at:
        type: string
      id:
        type: string
      last_error:
        type: string
      last_fetched_at:
        type: string
      name:
        type: string
      project_id:
        type: string
      status:
        $ref: '#/definitions/domain.SourceStatus'
      type:
        $ref: '#/definitions/domain.SourceType'
      updated_at:
        type: string
      url:
        type: string
    type: object
  domain.SourceStatus:
    enum:
    - active
    - error
    - disabled
    type: string
    x-enum-varnames:
    - SourceStatusActive
    - SourceStatusError
    - SourceStatusDisabled
  domain.SourceType:
    enum:
    - rss
    - web
    - telegram
    type: string
    x-enum-varnames:
    - SourceTypeRSS
    - SourceTypeWeb
    - SourceTypeTelegram
...
  handler.createSourceRequest:
    properties:
      config:
        items:
          type: integer
        type: array
      name:
        maxLength: 255
        minLength: 1
        type: string
      type:
        enum:
        - rss
        - web
        - telegram
        type: string
      url:
        type: string
    required:
    - name
    - type
    - url
    type: object
...
  handler.updateSourceRequest:
    properties:
      config:
        items:
          type: integer
        type: array
      name:
        maxLength: 255
        minLength: 1
        type: string
      url:
        type: string
    required:
    - name
    - url
    type: object
...
  httputil.ErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Brand reputation monitoring system API
  title: BrandRadar API
  version: "1.0"
paths:
...
  /projects/{projectID}/sources:
    get:
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Source'
            type: array
      security:
      - BearerAuth: []
      summary: List project sources
      tags:
      - sources
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Source data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createSourceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Source'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a source
      tags:
      - sources
  /projects/{projectID}/sources/{sourceID}:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Source ID
        in: path
        name: sourceID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a source
      tags:
      - sources
    put:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Source ID
        in: path
        name: sourceID
        required: true
        type: string
      - description: Source data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.updateSourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Source'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a source
      tags:
      - sources
  /projects/{projectID}/sources/{sourceID}/toggle:
    post:
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Source ID
        in: path
        name: sourceID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Source'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Toggle source status
      tags:
      - sources
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"